attention_layer/requirements.txt
--pre

# IREE
-f https://iree-org.github.io/iree/pip-release-links.html
iree-compiler
iree-runtime

# Torch-MLIR
-f https://llvm.github.io/torch-mlir/package-index/
torch-mlir

# PyTorch (CUDA)
--index-url https://download.pytorch.org/whl/nightly/cu117
torch
